# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/entrada": {
            "get": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Lista entradas do mês atual com paginação",
                "operationId": "9151f16c49de230a5e01bb1d4a184a83",
                "responses": {
                    "200": {
                        "description": "Lista de entradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AvailableMoney"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Cria nova entrada",
                "operationId": "6e750a545b843b39a06a857a306be5b4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "to_spend",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salário"
                                    },
                                    "to_spend": {
                                        "type": "number",
                                        "example": 2500
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entrada adicionada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableMoney"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/entrada/{id}": {
            "get": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Mostra uma entrada específica",
                "operationId": "a3462825de1a0efed4670def439d44ee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrada encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entrada não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Atualiza uma entrada existente",
                "operationId": "943e70fb1f1c581a52666c4b85b595ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "to_spend",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Salário Atualizado"
                                    },
                                    "to_spend": {
                                        "type": "number",
                                        "example": 3000
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-28"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Entrada atualizada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AvailableMoney"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Entrada não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Deleta uma entrada",
                "operationId": "6b4bf794d3702644dd4996610a4d43ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entrada deletada com sucesso"
                    },
                    "404": {
                        "description": "Entrada não encontrada"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/entrada/search": {
            "get": {
                "tags": [
                    "AvailableMoney"
                ],
                "summary": "Busca entradas por nome ou valor",
                "operationId": "d97708aef7226530755d856c810a197f",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Salário"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados da busca",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AvailableMoney"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AvailableMoney": {
                "title": "AvailableMoney",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "to_spend",
                    "date"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Saldo inicial"
                    },
                    "to_spend": {
                        "type": "number",
                        "format": "float",
                        "example": 1000.5
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-28"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-28T10:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Use o token de autenticação no formato Bearer <token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}